Add files via upload
Project 2- Business Statistics: Axis InsuranceLeveraging customer information is of paramount importance for most businesses. In the case of an insurance company, the attributes of customers can be crucial in making business decisions. Hence, knowing to explore and generate value out of such data can be an invaluable skill to have. EDA AND Statistical Analysis of Business Data was performed to achieve the following. 
Explore the dataset and extract insights using Exploratory Data Analysis.
Prove (or disprove) that the medical claims made by the people who smoke is greater than those who don't? [Hint- Formulate a hypothesis and prove/disprove it]
Prove (or disprove) with statistical evidence that the BMI of females is different from that of males.
Is the proportion of smokers significantly different across different regions?